More modules, add Dict support for serialiser
